Weekly task 4
Write a program that asks the user to input any positive integer and outputs the successive values of the following calculation. At each step calculate the next value by taking the current value and, if it is even, divide it by two, but if it is odd, multiply it by three and add one. Have the program end if the current value is one.
$ python collatz.py
Please enter a positive integer: 10
10 5 16 8 4 2 1

#ask user to enter a positive value (pedir al usuario que ingrese un valor positivo)
a = int(input("please enter a positive integer: "))

#if value entered is negative print "not a positive number" and stop (si el valor introducido es una impresion negativa "no es un número positivo" y se detiene)
while a < 0:
    print (a,"is not a positive number")
    break

#prints value of (a) when positive integer (imprime el valor de (a) cuando entero positivo)
while a > 0:
    print (a)
    break

#perform calculations until reaches 1 (realizar cálculos hasta llegar a 1)

while a > 1:

    if a % 2 == 0:
        a =  int(a/2)
        
    else:
        a = int((a*3)+1)

    print (a)